Move warningBoxStyle
