+++ src/pages/workspaces/workspace/DeleteWorkspaceModal.js
@@ -3,12 +3,12 @@ import pluralize from 'pluralize';
 import { useState } from 'react';
 import { b, div, h, label, li, p, span, strong, ul } from 'react-hyperscript-helpers';
 import { ButtonPrimary, Link, spinnerOverlay } from 'src/components/common';
-import { warningBoxStyle } from 'src/components/data/data-utils';
 import { icon } from 'src/components/icons';
 import { TextInput } from 'src/components/input';
 import Modal from 'src/components/Modal';
 import { bucketBrowserUrl } from 'src/libs/auth';
 import colors from 'src/libs/colors';
+import { warningBoxStyle } from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
 import { isAzureWorkspace, isGoogleWorkspace } from 'src/libs/workspace-utils';
 import { useDeleteWorkspaceState } from 'src/pages/workspaces/workspace/useDeleteWorkspaceState';
