+++ src/pages/workspaces/workspace/workflows/LaunchAnalysisModal.js
@@ -2,7 +2,6 @@ import _ from 'lodash/fp';
 import { Fragment, useState } from 'react';
 import { b, div, h, label, p, span } from 'react-hyperscript-helpers';
 import { ButtonPrimary, IdContainer, Link } from 'src/components/common';
-import { warningBoxStyle } from 'src/components/data/data-utils';
 import { icon, spinner } from 'src/components/icons';
 import { ValidatedTextArea } from 'src/components/input';
 import Modal from 'src/components/Modal';
@@ -13,6 +12,7 @@ import { launch } from 'src/libs/analysis';
 import colors from 'src/libs/colors';
 import { withErrorReporting } from 'src/libs/error';
 import { useCancellation, useOnMount } from 'src/libs/react-utils';
+import { warningBoxStyle } from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
 import { commentValidation } from 'src/pages/workspaces/workspace/jobHistory/UpdateUserCommentModal';
 import { chooseBaseType, chooseRootType, chooseSetType, processSnapshotTable } from 'src/pages/workspaces/workspace/workflows/EntitySelectionType';
